Move virtctl memorydump function to a separate file
The content is already enough to justify a new file
and in future commit we will add even more content.
