+++ pkg/virtctl/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
         "//pkg/virtctl/expose:go_default_library",
         "//pkg/virtctl/guestfs:go_default_library",
         "//pkg/virtctl/imageupload:go_default_library",
+        "//pkg/virtctl/memorydump:go_default_library",
         "//pkg/virtctl/pause:go_default_library",
         "//pkg/virtctl/portforward:go_default_library",
         "//pkg/virtctl/scp:go_default_library",
