+++ pkg/virtctl/memorydump/memorydump_suite_test.go
@@ -1,0 +1,11 @@
+package memorydump
+
+import (
+	"testing"
+
+	"kubevirt.io/client-go/testutils"
+)
+
+func TestVm(t *testing.T) {
+	testutils.KubeVirtTestSuiteSetup(t)
+}
