+++ pkg/virtctl/vm/vm.go
@@ -26,8 +26,6 @@ import (
 	"strings"
 
 	k8sv1 "k8s.io/api/core/v1"
-	k8serrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	v1 "kubevirt.io/api/core/v1"
@@ -37,8 +35,6 @@ import (
 
 	"kubevirt.io/client-go/kubecli"
 
-	storagetypes "kubevirt.io/kubevirt/pkg/storage/types"
-	kutil "kubevirt.io/kubevirt/pkg/util"
 	"kubevirt.io/kubevirt/pkg/virtctl/templates"
 )
 
@@ -51,25 +47,20 @@ const (
 	COMMAND_GUESTOSINFO    = "guestosinfo"
 	COMMAND_USERLIST       = "userlist"
 	COMMAND_FSLIST         = "fslist"
-	COMMAND_MEMORYDUMP     = "memory-dump"
 	COMMAND_ADDVOLUME      = "addvolume"
 	COMMAND_REMOVEVOLUME   = "removevolume"
 
 	volumeNameArg         = "volume-name"
-	claimNameArg          = "claim-name"
 	notDefinedGracePeriod = -1
 	dryRunCommandUsage    = "--dry-run=false: Flag used to set whether to perform a dry run or not. If true the command will be executed without performing any changes."
 
-	dryRunArg       = "dry-run"
-	pausedArg       = "paused"
-	forceArg        = "force"
-	gracePeriodArg  = "grace-period"
-	serialArg       = "serial"
-	persistArg      = "persist"
-	createClaimArg  = "create-claim"
-	storageClassArg = "storage-class"
-	accessModeArg   = "access-mode"
-	cacheArg        = "cache"
+	dryRunArg      = "dry-run"
+	pausedArg      = "paused"
+	forceArg       = "force"
+	gracePeriodArg = "grace-period"
+	serialArg      = "serial"
+	persistArg     = "persist"
+	cacheArg       = "cache"
 )
 
 var (
@@ -80,10 +71,6 @@ var (
 	persist      bool
 	startPaused  bool
 	dryRun       bool
-	claimName    string
-	createClaim  bool
-	storageClass string
-	accessMode   string
 	cache        string
 )
 
@@ -217,26 +204,6 @@ func NewFSListCommand(clientConfig clientcmd.ClientConfig) *cobra.Command {
 	return cmd
 }
 
-func NewMemoryDumpCommand(clientConfig clientcmd.ClientConfig) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "memory-dump get/remove (VM)",
-		Short:   "Dump the memory of a running VM to a pvc",
-		Example: usageMemoryDump(),
-		Args:    templates.ExactArgs("memory-dump", 2),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			c := Command{command: COMMAND_MEMORYDUMP, clientConfig: clientConfig}
-			return c.Run(args)
-		},
-	}
-	cmd.SetUsageTemplate(templates.UsageTemplate())
-	cmd.Flags().StringVar(&claimName, claimNameArg, "", "pvc name to contain the memory dump")
-	cmd.Flags().BoolVar(&createClaim, createClaimArg, false, "Create the pvc that will conatin the memory dump")
-	cmd.Flags().StringVar(&storageClass, storageClassArg, "", "The storage class for the PVC.")
-	cmd.Flags().StringVar(&accessMode, accessModeArg, "", "The access mode for the PVC.")
-
-	return cmd
-}
-
 func NewAddVolumeCommand(clientConfig clientcmd.ClientConfig) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "addvolume VMI",
@@ -319,22 +286,6 @@ func usage(cmd string) string {
 
 	usage := fmt.Sprintf("  # %s a virtual machine called 'myvm':\n", strings.Title(cmd))
 	usage += fmt.Sprintf("  {{ProgramName}} %s myvm", cmd)
-	return usage
-}
-
-func usageMemoryDump() string {
-	usage := `  #Dump memory of a virtual machine instance called 'myvm' to an existing pvc called 'memoryvolume'.
-  {{ProgramName}} memory-dump get myvm --claim-name=memoryvolume
-
-  #Create a PVC called 'memoryvolume' and dump the memory of a virtual machine instance called 'myvm' to it.
-  {{ProgramName}} memory-dump get myvm --claim-name=memoryvolume --create-claim
-
-  #Dump memory again to the same virtual machine with an already associated pvc(existing memory dump on vm status).
-  {{ProgramName}} memory-dump get myvm
-
-  #Remove the association of the memory dump pvc.
-  {{ProgramName}} memory-dump remove myvm
-  `
 	return usage
 }
 
@@ -424,118 +375,6 @@ func removeVolume(vmiName, volumeName, namespace string, virtClient kubecli.Kube
 		return fmt.Errorf("error removing volume, %v", err)
 	}
 	fmt.Printf("Successfully submitted remove volume request to VM %s for volume %s\n", vmiName, volumeName)
-	return nil
-}
-
-func calcMemoryDumpExpectedSize(vmName, namespace string, virtClient kubecli.KubevirtClient) (*resource.Quantity, error) {
-	vmi, err := virtClient.VirtualMachineInstance(namespace).Get(vmName, &metav1.GetOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	return kutil.CalcExpectedMemoryDumpSize(vmi), nil
-}
-
-func calcPVCNeededSize(memoryDumpExpectedSize *resource.Quantity, storageClass *string, virtClient kubecli.KubevirtClient) (*resource.Quantity, error) {
-	cdiConfig, err := virtClient.CdiClient().CdiV1beta1().CDIConfigs().Get(context.Background(), storagetypes.ConfigName, metav1.GetOptions{})
-	if k8serrors.IsNotFound(err) {
-		// can't properly determine the overhead - continue with default overhead of 5.5%
-		fmt.Printf(storagetypes.FSOverheadMsg)
-		return storagetypes.GetSizeIncludingDefaultFSOverhead(memoryDumpExpectedSize)
-	}
-	if err != nil {
-		return nil, err
-	}
-
-	fsVolumeMode := k8sv1.PersistentVolumeFilesystem
-	if *storageClass == "" {
-		storageClass = nil
-	}
-
-	return storagetypes.GetSizeIncludingFSOverhead(memoryDumpExpectedSize, storageClass, &fsVolumeMode, cdiConfig)
-}
-
-func createPVCforMemoryDump(namespace, vmName, claimName string, virtClient kubecli.KubevirtClient) error {
-	_, err := virtClient.CoreV1().PersistentVolumeClaims(namespace).Get(context.Background(), claimName, metav1.GetOptions{})
-	if err == nil {
-		return fmt.Errorf("PVC %s/%s already exists, check if it should be created if not remove create flag\n", namespace, claimName)
-	}
-	if !k8serrors.IsNotFound(err) {
-		return err
-	}
-
-	memoryDumpExpectedSize, err := calcMemoryDumpExpectedSize(vmName, namespace, virtClient)
-	if err != nil {
-		return err
-	}
-
-	neededSize, err := calcPVCNeededSize(memoryDumpExpectedSize, &storageClass, virtClient)
-	if err != nil {
-		return err
-	}
-
-	pvc := storagetypes.RenderPVC(neededSize, claimName, namespace, storageClass, accessMode, false)
-
-	_, err = virtClient.CoreV1().PersistentVolumeClaims(namespace).Create(context.Background(), pvc, metav1.CreateOptions{})
-	if err != nil {
-		return err
-	}
-
-	fmt.Printf("PVC %s/%s created\n", namespace, claimName)
-
-	return nil
-}
-
-func checkNoAssociatedMemoryDump(namespace, vmName string, virtClient kubecli.KubevirtClient) error {
-	vm, err := virtClient.VirtualMachine(namespace).Get(vmName, &metav1.GetOptions{})
-	if err != nil {
-		return err
-	}
-	if vm.Status.MemoryDumpRequest != nil {
-		return fmt.Errorf("please remove current memory dump association before creating a new claim for a new memory dump")
-	}
-	return nil
-}
-
-func memoryDump(args []string, claimName, namespace string, virtClient kubecli.KubevirtClient) error {
-	vmName := args[1]
-	switch args[0] {
-	case "get":
-		if createClaim {
-			if claimName == "" {
-				return fmt.Errorf("missing claim name")
-			}
-			if accessMode == string(k8sv1.ReadOnlyMany) {
-				return fmt.Errorf("cannot dump memory to a readonly pvc, use either ReadWriteOnce or ReadWriteMany if supported")
-			}
-			err := checkNoAssociatedMemoryDump(namespace, vmName, virtClient)
-			if err != nil {
-				return err
-			}
-			err = createPVCforMemoryDump(namespace, vmName, claimName, virtClient)
-			if err != nil {
-				return err
-			}
-		}
-		memoryDumpRequest := &v1.VirtualMachineMemoryDumpRequest{
-			ClaimName: claimName,
-		}
-
-		err := virtClient.VirtualMachine(namespace).MemoryDump(vmName, memoryDumpRequest)
-		if err != nil {
-			return fmt.Errorf("error dumping vm memory, %v", err)
-		}
-		fmt.Printf("Successfully submitted memory dump request of VM %s\n", vmName)
-	case "remove":
-		err := virtClient.VirtualMachine(namespace).RemoveMemoryDump(vmName)
-		if err != nil {
-			return fmt.Errorf("error removing memory dump association, %v", err)
-		}
-		fmt.Printf("Successfully submitted remove memory dump association of VM %s\n", vmName)
-	default:
-		return fmt.Errorf("invalid action type %s", args[0])
-	}
-
 	return nil
 }
 
@@ -689,8 +528,6 @@ func (o *Command) Run(args []string) error {
 		return addVolume(args[0], volumeName, namespace, virtClient, &dryRunOption)
 	case COMMAND_REMOVEVOLUME:
 		return removeVolume(args[0], volumeName, namespace, virtClient, &dryRunOption)
-	case COMMAND_MEMORYDUMP:
-		return memoryDump(args, claimName, namespace, virtClient)
 	}
 
 	fmt.Printf("VM %s was scheduled to %s\n", vmiName, o.command)
