+++ pkg/virtctl/vm/vm_test.go
@@ -3,7 +3,6 @@ package vm_test
 import (
 	"context"
 	"fmt"
-	"sync"
 
 	"k8s.io/utils/pointer"
 
@@ -13,21 +12,15 @@ import (
 	"github.com/spf13/cobra"
 
 	k8sv1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	k8smetav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 
 	v1 "kubevirt.io/api/core/v1"
 	"kubevirt.io/client-go/kubecli"
 	"kubevirt.io/containerized-data-importer-api/pkg/apis/core/v1beta1"
 
-	storagetypes "kubevirt.io/kubevirt/pkg/storage/types"
-	"kubevirt.io/kubevirt/pkg/virtctl/utils"
 	"kubevirt.io/kubevirt/tests/clientcmd"
 
 	"k8s.io/client-go/kubernetes/fake"
-	"k8s.io/client-go/testing"
 
 	cdifake "kubevirt.io/client-go/generated/containerized-data-importer/clientset/versioned/fake"
 )
@@ -571,279 +564,5 @@ var _ = Describe("VirtualMachine", func() {
 			Entry("with default", false),
 			Entry("with dry-run arg", true),
 		)
-	})
-
-	Context("memory dump", func() {
-		const (
-			createClaimFlag = "--create-claim"
-			claimNameFlag   = "--claim-name=testpvc"
-			claimName       = "testpvc"
-			configName      = "config"
-		)
-		var (
-			cdiClient       *cdifake.Clientset
-			coreClient      *fake.Clientset
-			pvcCreateCalled = &utils.AtomicBool{Lock: &sync.Mutex{}}
-		)
-
-		cdiConfigInit := func() (cdiConfig *v1beta1.CDIConfig) {
-			cdiConfig = &v1beta1.CDIConfig{
-				ObjectMeta: k8smetav1.ObjectMeta{
-					Name: storagetypes.ConfigName,
-				},
-				Spec: v1beta1.CDIConfigSpec{
-					UploadProxyURLOverride: nil,
-				},
-				Status: v1beta1.CDIConfigStatus{
-					FilesystemOverhead: &v1beta1.FilesystemOverhead{
-						Global: storagetypes.DefaultFSOverhead,
-					},
-				},
-			}
-			return
-		}
-		expectGetCDIConfig := func() {
-			kubecli.MockKubevirtClientInstance.EXPECT().CdiClient().Return(cdiClient).AnyTimes()
-		}
-
-		updateCDIConfig := func() {
-			config, err := cdiClient.CdiV1beta1().CDIConfigs().Get(context.Background(), configName, k8smetav1.GetOptions{})
-			Expect(err).ToNot(HaveOccurred())
-			config.Status.FilesystemOverhead.StorageClass = make(map[string]v1beta1.Percent)
-			config.Status.FilesystemOverhead.StorageClass["fakeSC"] = v1beta1.Percent("0.1")
-			_, err = cdiClient.CdiV1beta1().CDIConfigs().Update(context.Background(), config, k8smetav1.UpdateOptions{})
-			Expect(err).ToNot(HaveOccurred())
-		}
-
-		BeforeEach(func() {
-			pvcCreateCalled.False()
-			coreClient = fake.NewSimpleClientset()
-			cdiConfig := cdiConfigInit()
-			cdiClient = cdifake.NewSimpleClientset(cdiConfig)
-			kubecli.MockKubevirtClientInstance.EXPECT().CoreV1().Return(coreClient.CoreV1()).AnyTimes()
-		})
-
-		expectVMEndpointMemoryDump := func(vmName, claimName string) {
-			kubecli.MockKubevirtClientInstance.
-				EXPECT().
-				VirtualMachine(k8smetav1.NamespaceDefault).
-				Return(vmInterface).
-				Times(1)
-			vmInterface.EXPECT().MemoryDump(vmName, gomock.Any()).DoAndReturn(func(arg0, arg1 interface{}) interface{} {
-				Expect(arg0.(string)).To(Equal(vmName))
-				Expect(arg1.(*v1.VirtualMachineMemoryDumpRequest).ClaimName).To(Equal(claimName))
-				return nil
-			})
-		}
-
-		expectVMEndpointRemoveMemoryDump := func(vmName string) {
-			kubecli.MockKubevirtClientInstance.
-				EXPECT().
-				VirtualMachine(k8smetav1.NamespaceDefault).
-				Return(vmInterface).
-				Times(1)
-			vmInterface.EXPECT().RemoveMemoryDump(vmName).DoAndReturn(func(arg0 interface{}) interface{} {
-				Expect(arg0.(string)).To(Equal(vmName))
-				return nil
-			})
-		}
-
-		expectGetVM := func(withAssociatedMemoryDump bool) {
-			vm := &v1.VirtualMachine{
-				Spec:   v1.VirtualMachineSpec{},
-				Status: v1.VirtualMachineStatus{},
-			}
-			if withAssociatedMemoryDump {
-				vm.Status.MemoryDumpRequest = &v1.VirtualMachineMemoryDumpRequest{}
-			}
-			kubecli.MockKubevirtClientInstance.EXPECT().VirtualMachine(k8smetav1.NamespaceDefault).Return(vmInterface).Times(1)
-			vmInterface.EXPECT().Get(vmName, gomock.Any()).Return(vm, nil).Times(1)
-		}
-
-		expectGetVMNoAssociatedMemoryDump := func() {
-			expectGetVM(false)
-		}
-
-		expectGetVMWithAssociatedMemoryDump := func() {
-			expectGetVM(true)
-		}
-
-		expectGetVMI := func() {
-			quantity, _ := resource.ParseQuantity("256Mi")
-			vmi := &v1.VirtualMachineInstance{
-				Spec: v1.VirtualMachineInstanceSpec{
-					Domain: v1.DomainSpec{
-						Resources: v1.ResourceRequirements{
-							Requests: k8sv1.ResourceList{
-								k8sv1.ResourceMemory: quantity,
-							},
-						},
-					},
-				},
-			}
-			kubecli.MockKubevirtClientInstance.
-				EXPECT().
-				VirtualMachineInstance(k8smetav1.NamespaceDefault).
-				Return(vmiInterface).
-				Times(1)
-
-			vmiInterface.EXPECT().Get(vmName, gomock.Any()).Return(vmi, nil).Times(1)
-		}
-
-		pvcSpec := func() *k8sv1.PersistentVolumeClaim {
-			return &k8sv1.PersistentVolumeClaim{
-				ObjectMeta: k8smetav1.ObjectMeta{
-					Name: claimName,
-				},
-			}
-		}
-
-		expectPVCCreate := func(claimName, storageclass, accessMode string) {
-			coreClient.Fake.PrependReactor("create", "persistentvolumeclaims", func(action testing.Action) (bool, runtime.Object, error) {
-				create, ok := action.(testing.CreateAction)
-				Expect(ok).To(BeTrue())
-
-				pvc, ok := create.GetObject().(*k8sv1.PersistentVolumeClaim)
-				Expect(ok).To(BeTrue())
-				Expect(pvc.Name).To(Equal(claimName))
-
-				if storageclass != "" {
-					Expect(*pvc.Spec.StorageClassName).To(Equal(storageclass))
-					// 392Mi = (256Mi(vmi memory size) + 100Mi (memory dump overhead)) * 10%fsoverhead for fake storage class rounded to MiB
-					quantity, _ := resource.ParseQuantity("376Mi")
-					Expect(pvc.Spec.Resources.Requests[k8sv1.ResourceStorage]).To(Equal(quantity))
-				} else {
-					// 376Mi = (256Mi(vmi memory size) + 100Mi (memory dump overhead)) * 5.5%fsoverhead rounded to MiB
-					quantity, _ := resource.ParseQuantity("376Mi")
-					Expect(pvc.Spec.Resources.Requests[k8sv1.ResourceStorage]).To(Equal(quantity))
-				}
-				if accessMode != "" {
-					Expect(pvc.Spec.AccessModes[0]).To(Equal(k8sv1.PersistentVolumeAccessMode(accessMode)))
-				}
-
-				Expect(pvcCreateCalled.IsTrue()).To(BeFalse())
-				pvcCreateCalled.True()
-
-				return false, nil, nil
-			})
-		}
-
-		expectPVCGet := func(pvc *k8sv1.PersistentVolumeClaim) {
-			coreClient.Fake.PrependReactor("get", "persistentvolumeclaims", func(action testing.Action) (bool, runtime.Object, error) {
-				get, ok := action.(testing.GetAction)
-				Expect(ok).To(BeTrue())
-				Expect(get.GetNamespace()).To(Equal(k8smetav1.NamespaceDefault))
-				Expect(get.GetName()).To(Equal(claimName))
-				if pvc == nil {
-					return true, nil, errors.NewNotFound(v1.Resource("persistentvolumeclaim"), claimName)
-				}
-				return true, pvc, nil
-			})
-		}
-
-		DescribeTable("should fail with missing required or invalid parameters", func(errorString string, args ...string) {
-			commandAndArgs := []string{"memory-dump"}
-			commandAndArgs = append(commandAndArgs, args...)
-			cmdAdd := clientcmd.NewRepeatableVirtctlCommand(commandAndArgs...)
-			res := cmdAdd()
-			Expect(res).To(HaveOccurred())
-			Expect(res.Error()).To(ContainSubstring(errorString))
-		},
-			Entry("memorydump no args", "argument validation failed"),
-			Entry("memorydump missing action arg", "argument validation failed", "testvm"),
-			Entry("memorydump missing vm name arg", "argument validation failed", "get"),
-			Entry("memorydump wrong action arg", "invalid action type create", "create", "testvm"),
-			Entry("memorydump name, invalid extra parameter", "unknown flag", "testvm", "--claim-name=blah", "--invalid=test"),
-		)
-
-		It("should call memory dump subresource", func() {
-			expectVMEndpointMemoryDump("testvm", claimName)
-			commandAndArgs := []string{"memory-dump", "get", "testvm", claimNameFlag}
-			cmd := clientcmd.NewVirtctlCommand(commandAndArgs...)
-			Expect(cmd.Execute()).To(Succeed())
-		})
-
-		It("should call memory dump subresource without claim-name no create", func() {
-			expectVMEndpointMemoryDump("testvm", "")
-			commandAndArgs := []string{"memory-dump", "get", "testvm"}
-			cmd := clientcmd.NewVirtctlCommand(commandAndArgs...)
-			Expect(cmd.Execute()).To(Succeed())
-		})
-
-		It("should fail call memory dump subresource without claim-name with create-claim", func() {
-			commandAndArgs := []string{"memory-dump", "get", "testvm", createClaimFlag}
-			cmd := clientcmd.NewVirtctlCommand(commandAndArgs...)
-			res := cmd.Execute()
-			Expect(res).To(HaveOccurred())
-			Expect(res.Error()).To(ContainSubstring("missing claim name"))
-		})
-
-		It("should fail call memory dump subresource with create-claim with already associated memory dump pvc", func() {
-			expectGetVMWithAssociatedMemoryDump()
-			commandAndArgs := []string{"memory-dump", "get", "testvm", claimNameFlag, createClaimFlag}
-			cmd := clientcmd.NewVirtctlCommand(commandAndArgs...)
-			res := cmd.Execute()
-			Expect(res).To(HaveOccurred())
-			Expect(res.Error()).To(ContainSubstring("please remove current memory dump"))
-		})
-
-		It("should fail call memory dump subresource with create-claim and existing pvc", func() {
-			expectGetVMNoAssociatedMemoryDump()
-			expectPVCGet(pvcSpec())
-			commandAndArgs := []string{"memory-dump", "get", "testvm", claimNameFlag, createClaimFlag}
-			cmd := clientcmd.NewVirtctlCommand(commandAndArgs...)
-			res := cmd.Execute()
-			Expect(res).To(HaveOccurred())
-			Expect(res.Error()).To(ContainSubstring("already exists"))
-		})
-
-		It("should fail call memory dump subresource with create-claim no vmi", func() {
-			expectGetVMNoAssociatedMemoryDump()
-			kubecli.MockKubevirtClientInstance.EXPECT().VirtualMachineInstance(k8smetav1.NamespaceDefault).Return(vmiInterface).Times(1)
-			vmiInterface.EXPECT().Get(vmName, &k8smetav1.GetOptions{}).Return(nil, errors.NewNotFound(v1.Resource("virtualmachineinstance"), vmName))
-			commandAndArgs := []string{"memory-dump", "get", "testvm", claimNameFlag, createClaimFlag}
-			cmd := clientcmd.NewVirtctlCommand(commandAndArgs...)
-			res := cmd.Execute()
-			Expect(res).To(HaveOccurred())
-			Expect(res.Error()).To(ContainSubstring("not found"))
-		})
-
-		It("should fail call memory dump subresource with readonly access mode", func() {
-			commandAndArgs := []string{"memory-dump", "get", "testvm", claimNameFlag, createClaimFlag, "--access-mode=ReadOnlyMany"}
-			cmd := clientcmd.NewVirtctlCommand(commandAndArgs...)
-			res := cmd.Execute()
-			Expect(res).To(HaveOccurred())
-			Expect(res.Error()).To(ContainSubstring("cannot dump memory to a readonly pvc"))
-		})
-
-		DescribeTable("should create pvc for memory dump and call subresource", func(storageclass, accessMode string) {
-			expectGetCDIConfig()
-			expectGetVMNoAssociatedMemoryDump()
-			expectGetVMI()
-			expectPVCCreate(claimName, storageclass, accessMode)
-			expectVMEndpointMemoryDump("testvm", claimName)
-			commandAndArgs := []string{"memory-dump", "get", "testvm", claimNameFlag, createClaimFlag}
-			if storageclass != "" {
-				updateCDIConfig()
-				commandAndArgs = append(commandAndArgs, fmt.Sprintf("--storage-class=%s", storageclass))
-			}
-			if accessMode != "" {
-				commandAndArgs = append(commandAndArgs, fmt.Sprintf("--access-mode=%s", accessMode))
-			}
-			cmd := clientcmd.NewVirtctlCommand(commandAndArgs...)
-			Expect(cmd.Execute()).To(Succeed())
-			Expect(pvcCreateCalled.IsTrue()).To(BeTrue())
-		},
-			Entry("no other flags", "", ""),
-			Entry("with storageclass flag", "local", ""),
-			Entry("with access mode flag", "", "ReadWriteOnce"),
-		)
-
-		It("should call remove memory dump subresource", func() {
-			expectVMEndpointRemoveMemoryDump("testvm")
-			commandAndArgs := []string{"memory-dump", "remove", "testvm"}
-			cmd := clientcmd.NewVirtctlCommand(commandAndArgs...)
-			Expect(cmd.Execute()).To(Succeed())
-		})
 	})
 })
